Added applications-getting queries and endpoints and refactored all paged and tabled queries

+++ src/Application/Common/IPagedRequest.cs
@@ -4,6 +4,6 @@ namespace Application.Common;
 
 public interface IPagedRequest<T> : IRequest<PagedResultDto<T>>
 {
-    public long Page { get; }
-    public long ItemsPerPage { get; }
+    public int Page { get; }
+    public int ItemsPerPage { get; }
 }
