+++ src/Application/Common/PagedResultDto.cs
@@ -2,8 +2,8 @@
 
 public class PagedResultDto<T>
 {
-    public long CurrentPage { get; set; }
-    public long TotalItems { get; set; }
+    public int CurrentPage { get; set; }
+    public int TotalItems { get; set; }
     public bool HasNextPage { get; set; }
     public IEnumerable<T> Results { get; set; } = null!;
 }
